+++ src/Domains/Brokers/ObjectResponses/BrokerIntegrationSdkResolvedItem.php
@@ -15,12 +15,6 @@ final class BrokerIntegrationSdkResolvedItem
 
     public ?string $connection_name = null;
 
-    public ?string $mt5_server = null;
-
-    public ?int $mt5_port = null;
-
-    public ?string $mt5_login = null;
-
     public ?string $platform_server = null;
 
     public ?int $platform_port = null;